refactor: factor out fixed queue common code in an abstract class
